+++ wiki/ReleaseNotes.md
@@ -3,6 +3,7 @@
 * added support for `code` and `reason` parameters for `close()`
 * added support for specifying `binary` option when sending messages
 * `onclose` callback is invoked with event containing `code` and `reason`
+* `onerror` callback is invoked with event containing `code` and `reason`
 
 # wscpp 0.1.0
 This is the initial release of wscpp.